refactor txsource to make more testable
