+++ src/Nethermind/Nethermind.Merge.AuRa.Test/ShutterTransactionLoaderTests.cs
@@ -29,10 +29,10 @@ namespace Nethermind.Merge.AuRa.Test;
 using G1 = Bls.P1;
 using G2 = Bls.P2;
 using EncryptedMessage = ShutterCrypto.EncryptedMessage;
-using SequencedTransaction = ShutterTxSource.SequencedTransaction;
+using SequencedTransaction = ShutterTransactionLoader.SequencedTransaction;
 using Dto = Shutter.Dto;
 
-class ShutterTxSourceTests
+class ShutterTransactionLoaderSourceTests
 {
     [Test]
     public void Can_decrypt_sequenced_transactions()
@@ -45,7 +45,7 @@ class ShutterTxSourceTests
         byte[] msg = Convert.FromHexString("f869820248849502f900825208943834a349678ef446bae07e2aeffc01054184af008203e880824fd3a001e44318458b1f279bf81aef969df1b9991944bf8b9d16fd1799ed5b0a7986faa058f572cce63aaff3326df9c902d338b0c416c8fb93109446d6aadd5a65d3d115");
 
         List<SequencedTransaction> sequencedTransactions = [];
-        List<Dto.Key> keys = [];
+        List<(byte[] IdentityPreimage, byte[] Key)> keys = [];
 
         for (int i = 0; i < 100; i++)
         {
@@ -54,30 +54,26 @@ class ShutterTxSourceTests
             rnd.NextBytes(identityPrefix);
             rnd.NextBytes(sigma);
 
-            (SequencedTransaction sequencedTransaction, Dto.Key key) = GenerateSequencedTransaction(msg, identityPrefix, sk, new(sigma));
+            (SequencedTransaction sequencedTransaction, (byte[], byte[]) key) = GenerateSequencedTransaction(msg, identityPrefix, sk, new(sigma));
             sequencedTransactions.Add(sequencedTransaction);
             keys.Add(key);
         }
 
         // decryption keys are sorted by preimage
-        keys.Sort((a, b) => Bytes.BytesComparer.Compare(a.Identity.ToByteArray(), b.Identity.ToByteArray()));
+        keys.Sort((a, b) => Bytes.BytesComparer.Compare(a.IdentityPreimage, b.Key));
 
-        ShutterTxSource shutterTxSource = new ShutterTxSource(
+        ShutterTransactionLoader txLoader = new ShutterTransactionLoader(
             Substitute.For<ILogFinder>(),
             Substitute.For<IFilterStore>(),
-            Substitute.For<ReadOnlyTxProcessingEnvFactory>(),
-            Substitute.For<IAbiEncoder>(),
             Substitute.For<IShutterConfig>(),
             Substitute.For<ISpecProvider>(),
             new EthereumEcdsa(chainId, logManager),
             Substitute.For<IReadOnlyBlockTree>(),
-            Substitute.For<ShutterEon>(),
-            new Dictionary<ulong, byte[]>() {},
             logManager
         );
 
-        Transaction[] txs = shutterTxSource.DecryptSequencedTransactions(sequencedTransactions, keys);
-        foreach(Transaction tx in txs)
+        Transaction[] txs = txLoader.DecryptSequencedTransactions(sequencedTransactions, keys);
+        foreach (Transaction tx in txs)
         {
             byte[] tmp = Rlp.Encode<Transaction>(tx).Bytes;
             Assert.That(Enumerable.SequenceEqual(tmp, msg));
@@ -89,8 +85,8 @@ class ShutterTxSourceTests
     {
         // TestBlockchain
     }
-    
-    private (SequencedTransaction, Dto.Key) GenerateSequencedTransaction(byte[] msg, byte[] identityPreimage, UInt256 sk, Bytes32 sigma)
+
+    private (SequencedTransaction, (byte[] IdentityPreimage, byte[] Key)) GenerateSequencedTransaction(byte[] msg, byte[] identityPreimage, UInt256 sk, Bytes32 sigma)
     {
         G1 identity = ShutterCrypto.ComputeIdentity(identityPreimage);
         G2 eonKey = G2.generator().mult(sk.ToLittleEndian());
@@ -108,12 +104,6 @@ class ShutterTxSourceTests
             IdentityPreimage = identityPreimage
         };
 
-        var decryptionKey = new Dto.Key()
-        {
-            Identity = ByteString.CopyFrom(identityPreimage),
-            Key_ = ByteString.CopyFrom(key.compress())
-        };
-
-        return (sequencedTransaction, decryptionKey);
+        return (sequencedTransaction, (identityPreimage, key.compress()));
     }
 }
